requirements-dev.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
djangorestframework==3.13.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
